backend/pyproject.toml
[project]
name = "waifu-tutor-backend"
version = "0.1.0"
description = "Waifu Tutor FastAPI backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.6",
  "uvicorn[standard]>=0.32.1",
  "sqlalchemy>=2.0.36",
  "pydantic-settings>=2.7.0",
  "python-multipart>=0.0.12",
  "apscheduler>=3.10.4",
  "qdrant-client>=1.12.1",
  "google-genai>=1.2.0",
  "PyPDF2>=3.0.1",
  "python-docx>=1.1.2",
  "httpx>=0.28.1",
  "email-validator>=2.2.0"
]

[dependency-groups]
dev = [
  "pytest>=8.3.4",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.8.4"
]

[tool.pytest.ini_options]
pythonpath = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
